Added images and margins to game boxes
All games now have a centered thumbnail that is a test image for now.

I have added a margin to the game boxes so that the title and text don't
start at the leftmost boundry of the box

I need to correct the back background so that iut extends to cover the game
boxes, the current teqnique seems to work okay for pure text contents.